+++ app/src/components/common/Tab/Tabs.stories.tsx
@@ -1,0 +1,339 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { Tab, Tabs } from '.';
+
+const meta = {
+  title: 'Common/Tab/Tabs',
+  component: Tabs,
+  tags: ['autodocs'],
+} satisfies Meta<typeof Tabs>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {
+  render: () => (
+    <Tabs>
+      <Tab
+        selected
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        일반
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+    </Tabs>
+  ),
+};
+
+export const ManyTabs: Story = {
+  render: () => (
+    <Tabs>
+      <Tab
+        selected
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        일반
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+      <Tab
+        selected={false}
+        onClick={() => {
+          /* pass */
+        }}
+      >
+        평가
+      </Tab>
+    </Tabs>
+  ),
+};
